BootUpReceiver: handle QUICKBOOT, move TaskerService helper utils
 * android.intent.action.QUICKBOOT_POWERON
